Learn: move models out of models folder
