+++ src-ts/tools/learn/services/learn-provider/index.ts
@@ -1,3 +1,2 @@
 export * from './learn-provider.context'
-export * from './models'
 export * from './providers'
