--- src-ts/tools/learn/services/learn-provider/models/index.ts
@@ -1,5 +1,0 @@
-export * from './learn-certification.model'
-export * from './learn-course.model'
-export * from './learn-lesson.model'
-export * from './learn-module.model'
-export * from './learn-my-certification.model'
