+++ src-ts/tools/learn/services/learn-provider/providers/certifications.provider.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useCallback, useEffect, useMemo, useState } from 'react'
 
 import certificationsJSON from '../../../../../assets/data/certifications.json'
-import { LearnCertification } from '../models'
+import { LearnCertification } from './learn-certification.model'
 
 export interface CertificationsProviderData {
     certifications: Array<LearnCertification>
