+++ src-ts/tools/learn/services/learn-provider/providers/courses.provider.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useCallback, useEffect, useMemo, useState } from 'react'
 
 import coursesJSON from '../../../../../assets/data/courses.json'
-import { LearnCourse } from '../models'
+import { LearnCourse } from './learn-course.model'
 
 export interface CoursesProviderData {
     course?: LearnCourse
