+++ src-ts/tools/learn/services/learn-provider/providers/index.ts
@@ -1,5 +1,15 @@
-export * from './certifications-provider'
-export * from './courses-provider'
-export * from './lesson-provider'
-export * from './my-certifications-provider'
-export * from './my-certification-progress-provider'
+export * from './certifications.provider'
+export * from './courses.provider'
+export * from './lesson.provider'
+export * from './my-certifications.provider'
+export * from './my-certification-progress.provider'
+export * from './learn-certification.model'
+export * from './learn-course-provider.model'
+export * from './learn-course.model'
+export * from './learn-lesson-meta.model'
+export * from './learn-lesson.model'
+export * from './learn-module-meta.model'
+export * from './learn-module.model'
+export * from './learn-my-certification-progress.model'
+export * from './learn-my-certification.model'
+export * from './learn-my-module-progress.model'
