+++ src-ts/tools/learn/services/learn-provider/providers/learn-course-provider.model.ts
