+++ src-ts/tools/learn/services/learn-provider/providers/learn-course.model.ts
