+++ src-ts/tools/learn/services/learn-provider/providers/learn-lesson-meta.model.ts
@@ -11,8 +11,3 @@ export interface LearnLessonMeta {
     }
     title: string
 }
-
-export interface LearnLesson {
-    dashedName: string
-    title: string
-}
