+++ src-ts/tools/learn/services/learn-provider/providers/learn-lesson.model.ts
@@ -1,0 +1,4 @@
+export interface LearnLesson {
+    dashedName: string
+    title: string
+}
