+++ src-ts/tools/learn/services/learn-provider/providers/learn-module-meta.model.ts
@@ -1,5 +1,3 @@
-import { LearnLesson } from './learn-lesson.model'
-
 export interface LearnModuleMeta {
     dashedName: string
     estimatedCompletionTime: {
@@ -9,10 +7,4 @@ export interface LearnModuleMeta {
     introCopy: Array<string>
     lessonCount: number
     name: string
-}
-
-export interface LearnModule {
-    key: string
-    lessons: Array<LearnLesson>
-    meta: LearnModuleMeta
 }
