+++ src-ts/tools/learn/services/learn-provider/providers/learn-module.model.ts
@@ -1,0 +1,8 @@
+import { LearnLesson } from './learn-lesson.model'
+import { LearnModuleMeta } from './learn-module-meta.model'
+
+export interface LearnModule {
+    key: string
+    lessons: Array<LearnLesson>
+    meta: LearnModuleMeta
+}
