+++ src-ts/tools/learn/services/learn-provider/providers/learn-my-certification-progress.model.ts
@@ -1,0 +1,9 @@
+import { LearnMyModuleProgress } from './learn-my-module-progress.model'
+
+export interface LearnMyCertificationProgress {
+    completed: number
+    completedDate?: string
+    currentLesson?: string
+    modules: Array<LearnMyModuleProgress>
+    status: 'in-progress'|'completed'
+}
