+++ src-ts/tools/learn/services/learn-provider/providers/learn-my-certification.model.ts
@@ -1,0 +1,6 @@
+import { LearnCertification } from './learn-certification.model'
+import { LearnMyCertificationProgress } from './learn-my-certification-progress.model'
+
+export interface LearnMyCertification extends LearnCertification {
+    progress: LearnMyCertificationProgress
+}
