+++ src-ts/tools/learn/services/learn-provider/providers/learn-my-module-progress.model.ts
@@ -1,0 +1,9 @@
+export interface LearnMyModuleProgress {
+    completedLessons: Array<{
+        completedDate?: string
+        dashedName: string
+        title: string
+    }>
+    lessonCount: number
+    module: string
+}
