+++ src-ts/tools/learn/services/learn-provider/providers/my-certifications.provider.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useCallback, useEffect, useMemo, useState } from 'react'
 
 import myCertificationsJSON from '../../../../../assets/data/my-certifications.json'
-import { LearnMyCertification } from '../models'
+import { LearnMyCertification } from './learn-my-certification.model'
 
 const myCertifications: Array<LearnMyCertification> = [...myCertificationsJSON.certifications] as Array<LearnMyCertification>
